+++ packages/edit-widgets/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   The `GUTENBERG_PHASE` environment variable has been renamed to `IS_GUTENBERG_PLUGIN` and is now a boolean ([#38202](https://github.com/WordPress/gutenberg/pull/38202)).
 
+### Bug Fix
+
+-   Removed unused `@wordpress/a11y`, `@wordpress/server-side-render`, `rememo` and `uuid` dependencies ([#38388](https://github.com/WordPress/gutenberg/pull/38388)).
+
 ## 3.2.0 (2022-01-27)
 
 ## 3.1.0 (2021-11-07)